few more fixes for clarity
